+++ CHANGELOG.md
@@ -9,6 +9,7 @@ nav_order: 1
 ## [MAJOR.MINOR.PATCH] - YYYY-MM-DD
 - Fix user config serialization with null values only
 - Fix incorrect state name spelling in Flink resource
+- Add `aiven_flink_application` resource
 
 ## [3.11.0] - 2023-01-09
 - Fix races in tests